Move xenian_bot to xenian/bot for cleaner project structure
