+++ MANIFEST.in
@@ -1,4 +1,4 @@
-recursive-include xenian_bot *
+recursive-include xenian.bot *
 include *.txt
 include *.rst
 include *.md
