+++ xenian/bot/commands/filters/download_mode.py
@@ -1,7 +1,7 @@
 from telegram import Message
 from telegram.ext import BaseFilter
 
-from xenian_bot.utils import data
+from xenian.bot.utils import data
 
 __all__ = ['download_mode_filter']
