+++ xenian/bot/settings.example.py
@@ -9,7 +9,7 @@ YANDEX_API_TOKEN = 'YOUR_YANDEX_API_TOKEN'
 ADMINS = ['@SOME_TELEGRAM_USERS', ]  # Users which can do admin tasks like /restart
 SUPPORTER = ['@SOME_TELEGRAM_USERS', ]  # Users which to contact fo support
 
-TEMPLATE_DIR = os.path.join(BASE_DIR, 'xenian_bot/commands/templates')
+TEMPLATE_DIR = os.path.join(BASE_DIR, 'xenian.bot/commands/templates')
 
 # More information about polling and webhooks can be found here:
 # https://github.com/python-telegram-bot/python-telegram-bot/wiki/Webhooks
@@ -25,7 +25,7 @@ MODE = {
 }
 
 UPLOADER = {
-    'uploader': 'xenian_bot.uploaders.ssh.SSHUploader',  # What uploader to use
+    'uploader': 'xenian.bot.uploaders.ssh.SSHUploader',  # What uploader to use
     'url': 'YOUR_DOMAIN_FILES_DIR',
     'configuration': {
         'host': 'YOUR_HOST_IP',
