+++ xenian/bot/uploaders/base.py
