+++ xenian/bot/utils/template.py
